Add: file size in db

+++ ConvaReload.DataAccess/Migrations/ApplicationContextModelSnapshot.cs
@@ -86,8 +86,8 @@ namespace ConvaReload.DataAccess.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<int>("Size")
-                        .HasColumnType("integer");
+                    b.Property<long>("Size")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("UpdatedAt")
                         .HasColumnType("timestamp with time zone");
